fix(JSAM-2): register frontend
