+++ JSA-Mawsitsit-Frontend/App.js
@@ -4,8 +4,8 @@ import {
   StyleSheet,
   View,
 } from 'react-native';
-// import Register from './components/Register/Register';
-import Main from './components/Login/Main';
+import Register from './components/Register/Register';
+// import Main from './components/Login/Main';
 
 const styles = StyleSheet.create({
   container: {
@@ -17,8 +17,8 @@ const styles = StyleSheet.create({
 // Component Definition
 const App = () => (
   <View style={styles.container}>
-    {/* <Register /> */}
-    <Main />
+    <Register />
+    {/* <Main /> */}
   </View>
 );
